requirements.txt
asgiref==3.3.4
braintree==4.9.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.2
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-currentuser==0.5.3
django-debug-toolbar==3.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.0.0
django-import-export==2.5.0
django-json-field==0.5.7
django-rest-knox==4.1.0
django-taggit==1.4.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
geographiclib==1.50
geopy==2.1.0
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.0
MarkupPy==1.14
MarkupSafe==2.0.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
protobuf==3.17.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sentry-sdk==1.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
taggit-selectize==2.10.0
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
